move conftest into one place with the plugin too
